+++ src/qt/utilitydialog.cpp
@@ -10,9 +10,9 @@
 #include "clientmodel.h"
 #include "guiutil.h"
 
-#include "clientversion.h"
 #include "init.h"
 #include "util.h"
+#include "version.h"
 
 #include <QLabel>
 #include <QRegExp>
